Move component PageHeader from 'components' to 'components/_shared'
